requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
inflection==0.5.1
packaging==25.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
